+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockPlantPile.java
@@ -2,5 +2,5 @@ package org.allaymc.api.block.interfaces;
 
 import org.allaymc.api.block.BlockBehavior;
 
-public interface BlockLeafLitterBehavior extends BlockBehavior {
+public interface BlockPlantPile extends BlockBehavior {
 }