+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
 - Implemented small and big flowers:
   - Small flowers: Dandelion, Poppy, Blue Orchid, Allium, Azure Bluet, Red Tulip, Orange Tulip, White Tulip, Pink Tulip, Oxeye Daisy, Cornflower, Lily of the Valley, Wither Rose, Torchflower, Closed Eyeblossom, Open Eyeblossom.
   - Big flowers: Sunflower, Lilac, Rose Bush, Peony, Pitcher Plant
+- Implemented Leaf Litter, Pink Petals, and Wildflowers.
 
 ### Changed